added support for reading in Mako template to BbQuiz.py
BbQuiz.py will run its config file through Mako before passing it to
yaml.load, so it is now possible to embed calculations directly in the
template.

Currently, there is no way to get Mako to ignore missing variables, so
the references to LaTeX labels must be wrapped in ${''}, i.e.
${'${lbl_linear_force}'} instead of ${lbl_linear_force}. The
HomeworkAssignment class has been updated to support this.