HomeworkAssignment rewrite
Rewrote the pyHomework.HomeworkAssignment class to use classes instead
of a plain dict for storing homework information. The new class is
cleaner and more pythonic, but more importantly, it is more extensible.
Planning on adding the ability to place instructions between questions,
which was not possible in the previous design.