added support for user-defined macro hooks.
If a file named 'macro.py' is found in the current working directly, the
functions defined in it will be used for macro expansion. So, to provide
a handler for a new macro (or override an existing handler), you just
need to create a function with the macro name that takes a list of args
and an options string. For example

def mymacro( args, opts ):
  ...
  build the string that will replace the macro
  ...
  return text

will be called when \mymacro is found. The args argument is a list of
arguments, while the opts argument is just a string.

examples:
macro text   arguments
\mymacro                            -> args = [], opts = ''
\mymacro{some text}                 -> args = ['some text'], opts = ''
\mymacro{some text}{some more text} -> args = ['some text', 'some more text'], opts = ''
\mymacro[option text]{some text}    -> args = ['some text', 'some more text'], opts = 'option text'